+++ src/content/panorama/scripts/custom_game/ingame_hero_builder.js
@@ -14,10 +14,19 @@ function showIngameBuilder() {
         var balanceMode = GameUI.AbilityCosts.balanceModeEnabled == 1 ? true : false;
 
         // Spawn the hero builder
-        heroBuilderPanel = $.CreatePanel('Panel', $('#heroBuilderDisplay'), '');
-        heroBuilderPanel.BLoadLayout('file://{resources}/layout/custom_game/game_setup/game_setup.xml', false, false);
+        try {
+            heroBuilderPanel = $.GetContextPanel().GetParent().GetParent().GetParent().FindChildTraverse("CustomUIContainer_GameSetup").GetChild(0).GetChild(0);
+            if (heroBuilderPanel != null) {
+                heroBuilderPanel.SetParent($('#heroBuilderDisplay'))
+            } else {
+                throw true;
+            }
+        } catch (err) {
+            heroBuilderPanel = $.CreatePanel('Panel', $('#heroBuilderDisplay'), '');
+            heroBuilderPanel.BLoadLayout('file://{resources}/layout/custom_game/game_setup/game_setup.xml', false, false);
+        }
+        heroBuilderPanel.visible = true;
         heroBuilderPanel.isIngameBuilder = true;
-
         util.blockMouseWheel(heroBuilderPanel);
 
         // Boot it into selection mode
@@ -53,6 +62,7 @@ function showIngameBuilder() {
 		
 		util.reviewOptionsChange();
     } else {
+        heroBuilderPanel.visible = true;
         $('#heroBuilderDisplay').visible = !$('#heroBuilderDisplay').visible;
     }