+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ## 7.0 - Development Version
 - You can now use '-bot mode' chat command to see what mode bots are in (early game or late game). '-bot switch' can be used to alternative between modes, but can only be used when cheats are enable or in single player. 
 - Added cooldown to chaos meteor OP
+- Added '-god' and '-regen' cheat commands
 
 ## 6.9 - Steam Version
 - You can enable cheats in a multiplayer games now if everbody types '-enablecheats'