+++ CHANGELOG.md
@@ -7,6 +7,8 @@ All notable changes to this project will be documented in this file.
 - Added cooldown to chaos meteor OP
 - Added '-god' and '-regen' cheat commands
 - Revamped the ingame cheat menu, it is now a "command menu" which will house all added user commands. It is now accessible in all games not just ones with cheats mode on.
+- Optimized the ingame hero builder menu, should no longer cause instability or lags. 
+- Added a ingame filter to see hero perk related abiliites. 
 
 ## 6.9 - Steam Version
 - You can enable cheats in a multiplayer games now if everbody types '-enablecheats'